atleas

+++ app/Http/Controllers/PubTableController.php
@@ -7,6 +7,7 @@ use Illuminate\Support\Facades\File;
 use App\PubTable;
 use App\PubTableFiles;
 use App\Publikasi;
+use App\SubjectTable;
 
 class PubTableController extends Controller
 {
@@ -15,18 +16,25 @@ class PubTableController extends Controller
 
     public function pubTableForm($pub_id){
         $pubTable=[];
+        $daftar_subject=SubjectTable::all();
         $publikasi_detail=Publikasi::where('pub_id',$pub_id)->first();
         return view('publikasi.pub_table._addTableForm')
+        ->with('daftar_subject',$daftar_subject)
         ->with('pub_detail',$publikasi_detail)
         ->with('pubTable',$pubTable);
 
     }
 
     public function pubTableFormUpdate(Request $request,$pub_id,$pubtable_id){
+
+        $daftar_subject=SubjectTable::all();
+        
         $publikasi_detail=Publikasi::where('pub_id',$pub_id)->first();
         $pubTable=PubTable::find($pubtable_id);
         $pubTableFiles=PubTableFiles::where('pub_table_id',$pubTable->id)->get();
+        //dd($pubtable_id);
         return view('publikasi.pub_table._addTableForm')
+        ->with('daftar_subject',$daftar_subject)
         ->with('pub_detail',$publikasi_detail)
         ->with('pubTable',$pubTable)
         ->with('pubTableFile',$pubTableFiles);
@@ -41,11 +49,13 @@ class PubTableController extends Controller
         //$pdfFile->move($pub_id,$nama_file);
 
         $pubTable= PubTable::create([
-            'pub_id'=>$pub_id,
+            'type'=>"publikasi",
+            'type_id'=>$pub_id,
             'title'=>$request->tableTitle,
             'hal_pdf_first'=>$request->hal_pdf_first,
             'hal_pdf_last'=>$request->hal_pdf_last,
             'bab_num'=>$request->babNumberForm,
+            'subject_id'=>$request->subject_id
             //'pdf'=>$pdfFile
         ]);
 
@@ -60,54 +70,62 @@ class PubTableController extends Controller
 
         {
 
+            if($pubTable->type=="publikasi"){
+                $file=$request->file('PdfFileTable');
 
+                $name =$pubTable->hal_pdf_first.'_'.$pubTable->title.'.'.$file->extension();
+    
+                $file->move(public_path().'/files/'.$pubTable->pub_id, $name);  
+    
+                //$files[] = $name;  
+    
+                PubTableFiles::create([
+                    'pub_table_id'=>$pubTable->id,
+                    'filename'=>$name,
+                    'filepath'=>'files/'.$pubTable->pub_id.'/'.$name,
+                    'type'=>'pdf'
+                ]);
+    
+    
+                //multiple file
+               /*foreach($request->file('PdfFileTable') as $key=>$file)
+    
+               {
+                //echo($key);
+                $num=(int)$key;
+                $num=$num+1;
+               
+                   $name =$pubTable->title.$num.'.'.$file->extension();
+    
+                   $file->move(public_path().'/files/'.$pubTable->pub_id, $name);  
+    
+                   $files[] = $name;  
+    
+               }
+    
+               foreach($files as  $key=>$file){
+    
+                   PubTableFiles::create([
+                    'pub_table_id'=>$pubTable->id,
+                    'filename'=>$file,
+                    'filepath'=>'files/'.$pubTable->pub_id.'/'.$file,
+                    'type'=>'pdf'
+                   ]);
+               }*/
+
+               return redirect()->back();
+            }
+            
+            elseif ($pubTable->type=="lainnya") {
+                # code...
+            }
             //one file
-            $file=$request->file('PdfFileTable');
-
-            $name =$pubTable->hal_pdf_first.'_'.$pubTable->title.'.'.$file->extension();
-
-            $file->move(public_path().'/files/'.$pubTable->pub_id, $name);  
-
-            //$files[] = $name;  
-
-            PubTableFiles::create([
-                'pub_table_id'=>$pubTable->id,
-                'filename'=>$name,
-                'filepath'=>'files/'.$pubTable->pub_id.'/'.$name,
-                'type'=>'pdf'
-            ]);
-
-
-            //multiple file
-           /*foreach($request->file('PdfFileTable') as $key=>$file)
-
-           {
-            //echo($key);
-            $num=(int)$key;
-            $num=$num+1;
            
-               $name =$pubTable->title.$num.'.'.$file->extension();
-
-               $file->move(public_path().'/files/'.$pubTable->pub_id, $name);  
-
-               $files[] = $name;  
-
-           }
-
-           foreach($files as  $key=>$file){
-
-               PubTableFiles::create([
-                'pub_table_id'=>$pubTable->id,
-                'filename'=>$file,
-                'filepath'=>'files/'.$pubTable->pub_id.'/'.$file,
-                'type'=>'pdf'
-               ]);
-           }*/
 
         }
 
 
-        return redirect()->back();
+       
 
     }
 
@@ -128,66 +146,74 @@ class PubTableController extends Controller
                 'hal_pdf_first'=>$request->hal_pdf_first,
                 'hal_pdf_last'=>$request->hal_pdf_last,
                 'bab_num'=>$request->babNumberForm,
+                'subject_id'=>$request->subject_id
             ]
         );
 
             if($request->hasfile('PdfFileTable'))
             {
-                $file=$request->file('PdfFileTable');
+                if($pubTable->type=='publikasi'){
+
+                    $file=$request->file('PdfFileTable');
                 
-                $name =$pubTable->hal_pdf_first.'_'.$pubTable->title.'.'.$file->extension();
-
-
-                //delete file lama
-                if(File::exists(public_path('/'.$tableFiles->filepath)))
-                {
-                 File::delete(public_path('/'.$tableFiles->filepath));
-                }
-
-                //update file baru
-                 $file->move(public_path().'/files/'.$pubTable->pub_id, $name);  
-                
-
-                 $pubTableFiles=PubTableFiles::where('pub_table_id',$pubTable->id)->delete();
-
-                 PubTableFiles::create([
-                    'pub_table_id'=>$pubTable->id,
-                    'filename'=>$name,
-                    'filepath'=>'files/'.$pubTable->pub_id.'/'.$name,
-                    'type'=>'pdf'
-                   ]);
-             
-              /* foreach($request->file('PdfFileTable') as $key=>$file)
+                    $name =$pubTable->hal_pdf_first.'_'.$pubTable->title.'.'.$file->extension();
     
-               {
-                //echo($key);
-                $num=(int)$key;
-                $num=$num+1;
-               
-                   $name =$pubTable->title.'.'.$file->extension();
-
-                   if(File::exists(public_path('/files/'.$pubTable->pub_id.'/'.$name)))
-                   {
-                    File::delete(public_path('/files/'.$pubTable->pub_id.'/'.$name));
+                    if(!empty($tableFiles)){
+                        if(File::exists(public_path('/'.$tableFiles->filepath)))
+                        {
+                         File::delete(public_path('/'.$tableFiles->filepath));
+                        }
+        
                     }
+                    //delete file lama
 
-                    $file->move(public_path().'/files/'.$pubTable->pub_id, $name);  
+
+                    //update file baru
+                     $file->move(public_path().'/files/'.$pubTable->pub_id, $name);  
+                    
     
-                    $files[] = $name; 
-                  
-               }
-
-
-               $pubTableFiles=PubTableFiles::where('pub_table_id',$pubTable->id)->delete();
-               foreach($files as  $key=>$file){
+                     $pubTableFiles=PubTableFiles::where('pub_table_id',$pubTable->id)->delete();
     
-                   PubTableFiles::create([
-                    'pub_table_id'=>$pubTable->id,
-                    'filename'=>$file,
-                    'filepath'=>'files/'.$pubTable->pub_id.'/'.$file,
-                    'type'=>'pdf'
-                   ]);
-               }*/
+                     PubTableFiles::create([
+                        'pub_table_id'=>$pubTable->id,
+                        'filename'=>$name,
+                        'filepath'=>'files/'.$pubTable->pub_id.'/'.$name,
+                        'type'=>'pdf'
+                       ]);
+                 
+                  /* foreach($request->file('PdfFileTable') as $key=>$file)
+        
+                   {
+                    //echo($key);
+                    $num=(int)$key;
+                    $num=$num+1;
+                   
+                       $name =$pubTable->title.'.'.$file->extension();
+    
+                       if(File::exists(public_path('/files/'.$pubTable->pub_id.'/'.$name)))
+                       {
+                        File::delete(public_path('/files/'.$pubTable->pub_id.'/'.$name));
+                        }
+    
+                        $file->move(public_path().'/files/'.$pubTable->pub_id, $name);  
+        
+                        $files[] = $name; 
+                      
+                   }
+    
+    
+                   $pubTableFiles=PubTableFiles::where('pub_table_id',$pubTable->id)->delete();
+                   foreach($files as  $key=>$file){
+        
+                       PubTableFiles::create([
+                        'pub_table_id'=>$pubTable->id,
+                        'filename'=>$file,
+                        'filepath'=>'files/'.$pubTable->pub_id.'/'.$file,
+                        'type'=>'pdf'
+                       ]);
+                   }*/
+                }
+               
             }
 
             return redirect()->back();
